refactor: fix various build warnings
- split write-file utils to avoid importing nodejs dependencies into frontend
- increase chunk size warning limit to 2000
- suppress deprecation warnings coming from external dependencies
