+++ package.json
@@ -134,11 +134,11 @@
     "update-commit-hash:win32": "tools/get_app_version_for_windows.bat",
     "generate-notice": "mkdirp notices && yarn licenses generate-disclaimer --production > notices/notices.txt && yarn node tools/generateNotices.mjs && ./tools/checkNotices.sh",
     "ship-linux": "yarn install-deps:linux && yarn build && tools/build_linux_release.sh",
-    "ship-win": "yarn install-deps:win32 && yarn build && electron-builder --win --x64 --publish never && mkdirp release/win && mv \"release/OpossumUI Setup 0.1.0.exe\" \"release/win/OpossumUI-for-win.exe\"",
-    "ship-mac:x64": "yarn install-deps:darwin:x64 && yarn build && electron-builder --mac --x64 --publish never && cd release/mac && zip -r -y -q 'OpossumUI-for-mac-intel.zip' 'OpossumUI.app'",
-    "ship-mac:arm64": "yarn install-deps:darwin:arm64 && yarn build && electron-builder --mac --arm64 --publish never && codesign --force --deep -s - release/mac-arm64/OpossumUI.app && cd release/mac-arm64 && zip -r -y -q 'OpossumUI-for-mac-arm64.zip' 'OpossumUI.app'",
+    "ship-win": "yarn install-deps:win32 && yarn build && NODE_OPTIONS=--no-deprecation electron-builder --win --x64 --publish never && mkdirp release/win && mv \"release/OpossumUI Setup 0.1.0.exe\" \"release/win/OpossumUI-for-win.exe\"",
+    "ship-mac:x64": "yarn install-deps:darwin:x64 && yarn build && NODE_OPTIONS=--no-deprecation electron-builder --mac --x64 --publish never && cd release/mac && zip -r -y -q 'OpossumUI-for-mac-intel.zip' 'OpossumUI.app'",
+    "ship-mac:arm64": "yarn install-deps:darwin:arm64 && yarn build && NODE_OPTIONS=--no-deprecation electron-builder --mac --arm64 --publish never && codesign --force --deep -s - release/mac-arm64/OpossumUI.app && cd release/mac-arm64 && zip -r -y -q 'OpossumUI-for-mac-arm64.zip' 'OpossumUI.app'",
     "ship": "yarn ship-linux && yarn ship-win && yarn ship-mac",
-    "ship:auto": "run-script-os",
+    "ship:auto": "cross-env NODE_OPTIONS=--no-deprecation run-script-os",
     "ship:auto:darwin": "yarn node tools/runScriptArch.mjs ship-mac",
     "ship:auto:linux": "yarn ship-linux",
     "ship:auto:win32": "yarn ship-win",
